requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.3.7
Flask-Cors==5.0.1
Flask-WTF==1.2.2
python-socketio==5.11.4
eventlet==0.36.1
